requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
py2neo==2021.2.4
jieba==0.42.1
markdown==3.5.1
PyMySQL==1.1.0
python-dotenv==1.0.0
requests==2.31.0
lxml==4.9.3
fake-useragent==1.4.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
fasttext-wheel==0.9.2
transformers==4.30.2
torch==2.0.1